requirements.txt
# Core LLM and AI Framework
langchain
langchain-core
langgraph
bytez

# LLM Configuration
python-dotenv

# Data Processing
numpy
pydantic

# Logging and Output Formatting
colorama
rich

# Type Hints and Utilities
typing-extensions
